+++ tests/testthat/test-dispRity.metric.R
@@ -333,64 +333,9 @@ test_that("diagonal", {
 test_that("ancestral.dist", {
 
     set.seed(1)
-    tree <- rtree(6)
-    expect_equal(
-        get.ancestors(7, tree),
-        7)
-    expect_equal(
-        get.ancestors(4, tree),
-        c(11,10,9,8,7))
-    expect_equal(
-        get.ancestors(4, tree, full = FALSE),
-        c(11))
-
-    set.seed(1)
-    matrix <- matrix(rnorm(90), 9, 10)
-    tree <- rtree(5)
-    tree$node.label <- paste0("n", 1:4)
-    rownames(matrix) <- c(tree$tip.label, tree$node.label)
-
-    test <- nodes.coordinates(matrix, tree, full = FALSE)
-    expect_is(test, "matrix")
-    expect_equal(rownames(test), c("n2", "n2", "n3", "n4", "n4", "n1", "n1", "n1", "n3"))
-
-    test <- nodes.coordinates(matrix, tree, full = TRUE)
-    expect_is(test[[1]], "matrix")
-    expect_is(test[[2]], "matrix")
-    expect_is(test[[3]], "matrix")
-
-    expect_equal(rownames(test[[1]]), c("n2", "n2", "n3", "n4", "n4", "n1", "n1", "n1", "n3"))
-    expect_equal(rownames(test[[2]]), c("n1", "n1", "n1", "n3", "n3", NA, NA, NA, "n1"))
-    expect_equal(rownames(test[[3]]), c(NA, NA, NA, "n1", "n1", NA, NA, NA, NA))
-
-    set.seed(1)
     matrix <- matrix(rnorm(90), 9, 10)
     tree <- rtree(5) ; tree$node.label <- paste0("n", 1:4)
     rownames(matrix) <- c(tree$tip.label, tree$node.label)
-
-    direct_anc_centroids <- nodes.coordinates(matrix, tree, full = FALSE)
-    all_anc_centroids <- nodes.coordinates(matrix, tree, full = TRUE)
-
-    ## Old version (deprecated!)
-
-    test1 <- ancestral.dist.deprecated(matrix, nodes.coords = direct_anc_centroids)
-    test2 <- ancestral.dist.deprecated(matrix, nodes.coords = all_anc_centroids)
-
-    expect_equal(test1[6], c("n1" = 0))
-    expect_equal(test2[6], c("n1" = 0))
-    expect_lt(test1[1], test2[1])
-    expect_equal(test1[7], test2[7])
-
-    ## Ancestral dist without node coordinates
-    test3 <- ancestral.dist.deprecated(matrix, tree = tree, full = TRUE)
-    expect_warning(test4 <- ancestral.dist.deprecated(matrix))
-    expect_equal(names(test3), c(tree$tip.label, tree$node.label))
-    expect_equal(names(test4), c(tree$tip.label, tree$node.label))
-    expect_equal(test4, centroids(matrix))
-    
-    ## Ancestral dist with fixed node coordinates
-    test5 <- ancestral.dist.deprecated(matrix, nodes.coords = rep(0, ncol(matrix)))
-    expect_equal(test5, centroids(matrix, 0))
     
     ## New version
     expect_equal_round(ancestral.dist(matrix, tree), c(t3 = 3.226549, t5 = 4.014473, t4 = 3.392561, t2 = 3.963446, t1 = 4.389945, n1 = 0.000000, n2 = 4.938898, n3 = 3.490736, n4 = 2.746763), 6)